Refactor authorization logic and user info handling
- Replaced `IsUserValid` method with `GetUserInfo` in `IAuthorizationService` to return detailed user information.
- Updated `AuthEndPoints` to utilize the new method for user validation and added user roles to claims.
- Modified `ProductsEndPoints` to enforce authorization on specific endpoints and allow anonymous access for adding products.
- Changed HTTP request examples to reflect new user credentials and token usage.
- Implemented `GetUserInfo` in `AuthorizationService` with asynchronous logic.
- Introduced `UserInfo` record to encapsulate user details.

+++ FirstWebApp/EndPoints/ProductsEndPoints.cs
@@ -7,18 +7,19 @@ public static class ProductsEndPoints
 
     public static void MapProducts(this WebApplication app)
     {
-        var productsGroup = app.MapGroup("products").WithTags("Products");
+        var productsGroup = app.MapGroup("products")
+            .RequireAuthorization()
+            .WithTags("Products");
 
-        productsGroup.MapGet("", GetAllProducts).WithName(nameof(GetAllProducts));
+        productsGroup.MapGet("", GetAllProducts).WithName(nameof(GetAllProducts)).AllowAnonymous();
 
         productsGroup.MapGet("{id}", GetProductById).WithName(nameof(GetProductById));
 
-        productsGroup.MapDelete("{id}", DeleteItem).WithName(nameof(DeleteItem))
-            .RequireAuthorization(AuthorizationPolicies.ADMIN_POLICY);
+        productsGroup.MapPut("{id}", UpdateProduct).WithName(nameof(UpdateProduct)); 
 
-        productsGroup.MapPut("{id}", UpdateProduct).WithName(nameof(UpdateProduct)).RequireAuthorization(); 
+        productsGroup.MapPost("", AddNewProduct).WithName(nameof(AddNewProduct));
 
-        productsGroup.MapPost("", AddNewProduct).WithName(nameof(AddNewProduct)).RequireAuthorization();
+        productsGroup.MapDelete("{id}", DeleteItem).WithName(nameof(DeleteItem)).RequireAuthorization(AuthorizationPolicies.ADMIN_POLICY);
     }
 
     static async Task<Ok<List<ProductDto>>> GetAllProducts(IProductsRepository repository, IMapper mapper)
